// mern-project/src/component/utils/types.enum.ts
/**
 * @description: Catergory.enum
 * 
 */
export enum StepStatus {
    error = 'error',
    process = 'process',
    finish = 'finish',
    wait = 'wait'
}


/**
 * @description: accessories.enum
 * 
 */
export enum RAM {
    DDR4_4 = "4GB",
    DDR4_8 = "8GB",
    DDR4_16 = "16GB",
    DDR4_32 = "32GB"
}
export enum SSD {
    PCIE_128 = "PCIE128",
    PCIE_256 = "PCIE256",
    PCIE_512 = "PCIE512",
    PCIE_1024 = "PCIE1024",
    PCIE_2048 = "PCIE2048"
}
export enum HDD {
    HDD_1TB = "1TB",
    HDD_2TB = "2TB",
    HDD_3TB = "3TB"
}
export enum OS {
    W10H = "W10H",
    W11H = "W11H",
    W10P = "W10P",
    W11P = "W11P"
}